+++ WrightTools/artists/__init__.py
@@ -6,4 +6,4 @@ from ._base import *
 from ._colors import *
 from ._helpers import *
 from ._quick import *
-from ._interactive import *
+from ._interact import *
